Huge changes to Move
I'm either a genius, or an idiot. So I've changed Move.java entirely where there are conditionals, that will determine how the move should be performed and what methods it should use. Also begun on the general outline of everything. Essentially I'm going for a complete revamp of the code