Changed Pokemon.java
Pokemon.java needed to be the object that will be created through the json files. Also to prevent problems with Jackson, I just set all of the basestats to public instead of private with getter/setter. I believe I've accounted for all instances where .get a basestat was used as well