Logic Change
Thank god it worked. I've started on the creation of the interface.java object. The reasoning is that, I want a place where all the modifications, player choices and such are stored. Not sure if later down the line I will keep this approach. But certainly it's a huge change. I believe this will also make the program smaller, since the Json files exist as attributes for Interface and much else as well.