hopefully works
I hope this works as intended cause that would be sick. But yeah I've used a bubble sort that checks for both if prio is greater than next element and prio is equal while the speed is greater then they'll swap. This way I have to ordering of the pokemon. I've converted under start battle everything that should be converted, so no more arrays of indexes, simply arrays of Pokemon and moves.