Rewriting interface
adding pokemon as well. Also a thing to note is that the Jackson library is getting overwhelmed by the poketiering attempt file, therefor, I need to use parsing and more stuff. So either I'll adjust how many pokemon I have or something else. Likewise I've converted everything under player to arrayList i.e. instead of having an ugly array, that needs to constantly be ready from and converted to an arrayList thereafter, I instead only have ArrayLists and ready from those directly. I've started on converting moves under pokemon to that as well, but sadly there are some problems with the if statement block reading PP as a conditional. All in all an alright day in terms of productivity.