Interface change
Since the pokemon dataset become to large, I needed to have a "lazy" way to look up all pokemon instantly. The most efficient is a hashMap where there's a String Pokemon name that maps to the pokemon itself. It's the same as before, but I simply load everything into the hashMap called allPokemonMapped, when constructing an Interface object. This is a better, more cost efficient approach, compared to before. And I can virtually have infinite pokemon now (I think). also simply using Pokemon somePokemon allPokemonMapped.get(nameOfAnyPokemon) is enough to get a pokemon directly. Naturally if and only if it's through an interface object that has used the method loadPokemonData.