.
a bit to add on. Likewise I didn't clarify that the reason I needed to move it up to Team level (the controller) was that I was thinking that I want to avoid recursion. Since sadly it gathers stuff up on the "heap", don't really understand the "heap" yet. But my basic understanding is that it's what the compiler is reading? At least having a clean stack is preferred and avoids bugs.